+++ src/apperror.rs
@@ -8,14 +8,12 @@ pub enum AppError {
     FileReadError,
     FileSkipError,
     FileWriteError,
-    FullFilenameError,
     FileNotFound,
     ProtectFile,
     FileDataError,
     SaveError,
     SystemTimeError,
     CannotCreateDir,
-    
 }
 
 impl std::error::Error for AppError {}
@@ -30,13 +28,12 @@ impl std::fmt::Display for AppError {
             Self::FileSkipError => "跳过数据时出错。可能是文件大小小于预期",
             Self::FileReadError => "读取文件时发生错误",
             Self::FileWriteError => "写入文件时错误",
-            Self::FullFilenameError => "文件名不符合规范",
             Self::FileNotFound => "未找到文件",
             Self::ProtectFile => "已关闭文件强制覆盖且文件已存在。使用-f或-forcesave开启强制覆盖。",
             Self::FileDataError => "处理文件数据时出错",
             Self::SaveError => "保存文件出错",
             Self::SystemTimeError => "获取时间戳失败",
-            Self::CannotCreateDir => "无法创建父级目录", // _ =>  "未知错误",
+            Self::CannotCreateDir => "无法创建父级目录",
         };
         write!(f, "{}", result)
     }